Rotation support, Calc Bug, TextView Init
Added onInstanceSave and onRestoreInstance so that the values are
protected during orientation changes and when leaving app onPause.
Declared TextView as a class field then gave it an id after set content
view in onCreate, do not have to findViewById multiple times, efficiency
improvement.  Fixed minor calculation error that occured on
initialization in which after selecting first operation first input was
set to zero.